update de clase deteccion

+++ scripts/detection_adjust.py
@@ -1,5 +1,5 @@
 # TODO: 'convertir todo esto a una clase con métodos:
-#  1) buscar los lower bounds mediantes los sliders y que la func retorne la
+#  --) buscar los lower bounds mediantes los sliders y que la func retorne la
 #     máscara con los valores de los threshold
 #  2) tomar esos valores y generar la imagen correspondiente con la firma con-
 #     vertida a solo la curva
@@ -11,6 +11,7 @@
 
 import cv2
 import numpy as np
+import matplotlib.pyplot as plt
 
 
 class Detector(object):
@@ -21,48 +22,51 @@ class Detector(object):
         self.mask = 0
         self.img = 0
         self.img_extracted = 0
+        self.original = 0
 
     def add_img(self, path_to_img):
-        initial = cv2.imread(path_to_img)
-        self.img = cv2.cvtColor(initial, cv2.COLOR_BGR2HSV)
-        self.img_extracted = self.img.copy()
+        self.original = cv2.imread(path_to_img)
+        self.img = cv2.cvtColor(self.original, cv2.COLOR_BGR2HSV)
+        self.img_extracted = self.original.copy()
 
     def nothing(self, x):
         pass
 
     def get_mask_bounds(self, path_to_img):
-        self.add_img(path_to_img)
-        cv2.namedWindow('image', cv2.WINDOW_AUTOSIZE)
-        cv2.createTrackbar('R', 'image', 0, 255, self.nothing)
-        cv2.createTrackbar('G', 'image', 0, 255, self.nothing)
-        cv2.createTrackbar('B', 'image', 0, 255, self.nothing)
-        cv2.createTrackbar('Ru', 'image', 0, 255, self.nothing)
-        cv2.createTrackbar('Gu', 'image', 0, 255, self.nothing)
-        cv2.createTrackbar('Bu', 'image', 0, 255, self.nothing)
+        mode = input('Set bound manually? [y/n]')
+        if mode == 'y':
+            self.add_img(path_to_img)
+            cv2.namedWindow('image', cv2.WINDOW_AUTOSIZE)
+            cv2.createTrackbar('R', 'image', 0, 255, self.nothing)
+            cv2.createTrackbar('G', 'image', 0, 255, self.nothing)
+            cv2.createTrackbar('B', 'image', 0, 255, self.nothing)
+            cv2.createTrackbar('Ru', 'image', 0, 255, self.nothing)
+            cv2.createTrackbar('Gu', 'image', 0, 255, self.nothing)
+            cv2.createTrackbar('Bu', 'image', 0, 255, self.nothing)
 
-        while ((cv2.waitKey(1) & 0xFF) != 27):
-            r = cv2.getTrackbarPos('R', 'image')
-            g = cv2.getTrackbarPos('G', 'image')
-            b = cv2.getTrackbarPos('B', 'image')
-            ru = cv2.getTrackbarPos('Ru', 'image')
-            gu = cv2.getTrackbarPos('Gu', 'image')
-            bu = cv2.getTrackbarPos('Bu', 'image')
-            lower_lim = np.array([r, g, b])
-            upper_lim = np.array([ru, gu, bu])
+            while ((cv2.waitKey(1) & 0xFF) != 27):
+                r = cv2.getTrackbarPos('R', 'image')
+                g = cv2.getTrackbarPos('G', 'image')
+                b = cv2.getTrackbarPos('B', 'image')
+                ru = cv2.getTrackbarPos('Ru', 'image')
+                gu = cv2.getTrackbarPos('Gu', 'image')
+                bu = cv2.getTrackbarPos('Bu', 'image')
+                lower_lim = np.array([r, g, b])
+                upper_lim = np.array([ru, gu, bu])
 
-            mask = cv2.inRange(self.img, lower_lim, upper_lim)
+                mask = cv2.inRange(self.img, lower_lim, upper_lim)
 
-            kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (3, 3))
-            opening = cv2.morphologyEx(mask, cv2.MORPH_OPEN, kernel, iterations=1)
-            close = cv2.morphologyEx(opening, cv2.MORPH_CLOSE, kernel, iterations=2)
+                kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (3, 3))
+                opening = cv2.morphologyEx(mask, cv2.MORPH_OPEN, kernel, iterations=1)
+                close = cv2.morphologyEx(opening, cv2.MORPH_CLOSE, kernel, iterations=2)
 
-            cv2.imshow('image', close)
+                cv2.imshow('image', close)
 
-        cv2.destroyAllWindows()
+            cv2.destroyAllWindows()
 
-        self.lower = lower_lim
-        self.upper = upper_lim
-        self.mask = close
+            self.lower = lower_lim
+            self.upper = upper_lim
+            self.mask = close
 
     def get_only_curve(self):
         cnts = cv2.findContours(self.mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
@@ -90,14 +94,34 @@ class Detector(object):
             # cv2.imwrite('result.png', result)
             # cv2.imwrite('ROI.png', ROI)
             cv2.waitKey()
+            #cv2.imwrite('mascar.png', self.mask)
 
         except:
             print('Not enough contours found')
 
 
+    def get_final_countours(self):
+        img = cv2.imread('mascar.png')
+        self.add_img('firma2.png')
+        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+        #contours = cv2.findContours(img, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)[-2]
+        contours, hierarchy = cv2.findContours(gray,
+                                               cv2.RETR_TREE,
+                                               cv2.CHAIN_APPROX_SIMPLE)
+
+        cv2.drawContours(self.original, contours, -1, (0, 0, 0), 2)
+        cv2.imshow('Contours', self.original)
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
+
+    def convert_to_svg(self):
+       pass
+
+
 
 if __name__ == "__main__":
     yay = Detector()
-    yay.get_mask_bounds('firma2.png')
-    print(yay.lower, yay.upper)
+    yay.get_mask_bounds('chionpaper.jpg')
+    #print(yay.lower, yay.upper)
     yay.get_only_curve()
+    #yay.get_final_countours()
